requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.4
drf-yasg>=1.20.0
gunicorn>=20.1.0
web3>=5.24.0
python-dotenv>=0.19.0